add getParsedFromLS och SetArrayOrRemove to LS and move createNewId to a dataHelpers file
