+++ src/services/DreamService.ts
@@ -1,9 +1,10 @@
 import { Dream, newDream } from "../models/types";
 import { dreams } from "../store/globalVariables.js";
-import { LS, createNewId, getParsedFromLS, setArrayOrRemove } from "../utils/localStorageHelpers.js"
+import { LS, } from "../utils/localStorageHelpers.js";
+import { createNewId } from "../utils/dataHelpers.js"
 
 export function getDreams(): Dream[] {
-    return getParsedFromLS<Dream[]>("dreams", dreams);
+    return LS.getParsedFromLS<Dream[]>("dreams", dreams);
 }
 
 export function addDream(newDream: newDream): void {
@@ -12,12 +13,12 @@ export function addDream(newDream: newDream): void {
     const updatedDream = { ...newDream, id: newId };
 
     currentDreams.push(updatedDream);
-    
+
     LS.add("dreams", JSON.stringify(currentDreams))
 }
 
 export function updateDream(updatedDream: Dream): void {
-    const currentDreams = getParsedFromLS<Dream[]>("dreams", dreams);
+    const currentDreams = LS.getParsedFromLS<Dream[]>("dreams", dreams);
     const updatedDreams = currentDreams.map(dream => dream.id === updatedDream.id ? updatedDream : dream);
 
     LS.add("dreams", JSON.stringify(updatedDreams))
@@ -25,8 +26,8 @@ export function updateDream(updatedDream: Dream): void {
 }
 
 export function deleteDream(id: number): void {
-    const currentDreams = getParsedFromLS<Dream[]>("dreams", dreams);
+    const currentDreams = LS.getParsedFromLS<Dream[]>("dreams", dreams);
     const updatedDreams = currentDreams.filter(dream => dream.id !== id);
 
-    setArrayOrRemove<Dream>("dreams",updatedDreams);
+    LS.setArrayOrRemove<Dream>("dreams", updatedDreams);
 }
