+++ src/services/ThemeService.ts
@@ -1,12 +1,14 @@
 import { Theme } from "../models/types";
 import { themes } from "../store/globalVariables.js";
-import { LS, createNewId, getParsedFromLS, setArrayOrRemove } from "../utils/localStorageHelpers.js"
+import { LS } from "../utils/localStorageHelpers.js";
+import { createNewId } from "../utils/dataHelpers.js"
+
 
 export function getThemes(): Theme[] {
-    return getParsedFromLS<Theme[]>("themes", themes);
+    return LS.getParsedFromLS<Theme[]>("themes", themes);
 }
 
- export function addTheme(themeName: string): void {
+export function addTheme(themeName: string): void {
     const currentThemes = getThemes();
     const newId = createNewId(currentThemes);
     const updatedTheme = { name: themeName, id: newId };
@@ -18,7 +20,7 @@ export function getThemes(): Theme[] {
 
 
 export function updateTheme(updatedTheme: Theme): void {
-    const currentThemes = getParsedFromLS<Theme[]>("themes", themes);
+    const currentThemes = LS.getParsedFromLS<Theme[]>("themes", themes);
     const updatedThemes = currentThemes.map(theme => theme.id === updatedTheme.id ? updatedTheme : theme);
 
     LS.add("themes", JSON.stringify(updatedThemes))
@@ -26,8 +28,8 @@ export function updateTheme(updatedTheme: Theme): void {
 }
 
 export function deleteTheme(id: number): void {
-    const currentThemes = getParsedFromLS<Theme[]>("themes", themes);
+    const currentThemes = LS.getParsedFromLS<Theme[]>("themes", themes);
     const updatedThemes = currentThemes.filter(theme => theme.id !== id);
 
-    setArrayOrRemove<Theme>("themes",updatedThemes);
+    LS.setArrayOrRemove<Theme>("themes", updatedThemes);
 }
