+++ src/utils/dataHelpers.ts
@@ -1,0 +1,3 @@
+export function createNewId<T extends { id: number }>(array: T[]): number {
+    return Math.max(0, ...array.map(item => item.id)) + 1;
+}
